Combine the tree viewer performance tests into one class
TreeAddTest, FastTreeTest, ShrinkingTreeTest and TreeViewerRefreshTest
each measured one aspect of the same widget in its own class, which the
old performance database needed because a class was the unit it reported
on. Now that the tests print their own timings the split only spreads
one subject over four files and a shared base class.

TreeViewerPerformanceTest carries all of the scenarios. The additions
that FastTreeTest repeated ten times with a removal in between become a
round count on the shared measuring method, and the two refresh
scenarios keep their content providers, so the measurements are the ones
the separate classes took. TreeTest is gone, its content provider now
sits with the test that uses it.
