+++ tests/org.eclipse.ui.tests.performance/src/org/eclipse/jface/tests/performance/JFacePerformanceSuite.java
@@ -24,11 +24,9 @@ import org.junit.platform.suite.api.SelectClasses;
 		ListViewerRefreshTest.class, //
 		ComboViewerRefreshTest.class, //
 		FastTableViewerRefreshTest.class, //
-		FastTreeTest.class, //
-		TreeAddTest.class, //
+		TreeViewerPerformanceTest.class, //
 		TreeViewerExpansionTest.class, //
 		ProgressMonitorDialogPerformanceTest.class, //
-		ShrinkingTreeTest.class, //
 		CollatorPerformanceTest.class, //
 })
 public class JFacePerformanceSuite {
