--- tests/org.eclipse.ui.tests.performance/src/org/eclipse/jface/tests/performance/TreeViewerRefreshTest.java
@@ -1,63 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2005, 2006 IBM Corporation and others.
- *
- * This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License 2.0
- * which accompanies this distribution, and is available at
- * https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     IBM Corporation - initial API and implementation
- *******************************************************************************/
-
-package org.eclipse.jface.tests.performance;
-
-import static org.eclipse.ui.tests.harness.util.UITestUtil.processEvents;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import org.eclipse.jface.viewers.StructuredViewer;
-import org.eclipse.jface.viewers.TreeViewer;
-import org.eclipse.swt.widgets.Shell;
-import org.junit.jupiter.api.Test;
-
-
-/**
- * The TreeViewerRefreshTest is the refresh test for
- * tree viewers.
- */
-public class TreeViewerRefreshTest extends ViewerTest {
-
-	TreeViewer viewer;
-	private RefreshTestTreeContentProvider contentProvider;
-
-	@Override
-	protected StructuredViewer createViewer(Shell shell) {
-		viewer = new TreeViewer(shell);
-		contentProvider = new RefreshTestTreeContentProvider();
-		viewer.setContentProvider(contentProvider);
-		viewer.setLabelProvider(getLabelProvider());
-		return viewer;
-	}
-
-	/**
-	 * Test the time for doing a refresh.
-	 */
-	@Test
-	public void testRefresh() throws Throwable {
-		openBrowser();
-
-		for (int i = 0; i < ITERATIONS; i++) {
-			startMeasuring();
-			viewer.refresh();
-			processEvents();
-			stopMeasuring();
-			assertEquals(RefreshTestContentProvider.ELEMENT_COUNT, viewer.getTree().getItemCount());
-		}
-
-		reportTimings();
-	}
-
-
-}
